requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
win32_setctime==1.2.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.2.1